using states to control flow of bot. Added a dropdown menu during item choosing process.
Still have some minor bugs, edit item during createloan process is not done. During choosing the quantity of the item, if user chooses the quantity of an item two times, it should add the quantity to the item instead of creating a newline for that item. ViewLoan command is not done too.